+++ R/options.R
@@ -55,7 +55,7 @@
 #'     A non-negative scalar on how intrusive (disruptive) the reporter to the user. This multiplicative scalar applies to the _interval_ and _times_ parameters. (Default: `1.0`)\cr
 #'   
 #'     \describe{
-#'       \item{\option{progressr.intrusiveness.auditory}:}{(numeric) intrusiveness for auditory progress handlers (Default: `5.0`)}
+#'       \item{\option{progressr.intrusiveness.audio}:}{(numeric) intrusiveness for auditory progress handlers (Default: `5.0`)}
 #'       \item{\option{progressr.intrusiveness.file}:}{(numeric) intrusiveness for file-based progress handlers (Default: `5.0`)}
 #'       \item{\option{progressr.intrusiveness.gui}:}{(numeric) intrusiveness for graphical-user-interface progress handlers (Default: `1.0`)}
 #'       \item{\option{progressr.intrusiveness.notifier}:}{(numeric) intrusiveness for progress handlers that creates notifications (Default: `10.0`)}
@@ -130,7 +130,7 @@
 #' progressr.interrupts
 #' progressr.interval
 #' progressr.intrusiveness
-#' progressr.intrusiveness.auditory
+#' progressr.intrusiveness.audio
 #' progressr.intrusiveness.debug
 #' progressr.intrusiveness.file
 #' progressr.intrusiveness.gui